rename and improve lab page
